:bug: #56 데모 페이지 서버 url 및 개인 정보 config로 분리
